onewire plugin: fix two compiler warnings
onewire.c:418:31: warning: comparison of integers of different signs:
'int' and 'unsigned long' [-Wsign-compare]
  if ((status > 0) && (status < sizeof (subpath)))
                       ~~~~~~ ^ ~~~~~~~~~~~~~~~~
onewire.c:422:31: warning: comparison of integers of different signs:
'int' and 'unsigned long' [-Wsign-compare]
  if ((status > 0) && (status < sizeof (subpath)))
                       ~~~~~~ ^ ~~~~~~~~~~~~~~~~

+++ src/onewire.c
@@ -415,11 +415,11 @@ static int cow_read_ds2409 (const char *path)
   int status;
 
   status = ssnprintf (subpath, sizeof (subpath), "%s/main", path);
-  if ((status > 0) && (status < sizeof (subpath)))
+  if ((status > 0) && (status < (int) sizeof (subpath)))
     cow_read_bus (subpath);
 
   status = ssnprintf (subpath, sizeof (subpath), "%s/aux", path);
-  if ((status > 0) && (status < sizeof (subpath)))
+  if ((status > 0) && (status < (int) sizeof (subpath)))
     cow_read_bus (subpath);
 
   return (0);
